requirements.txt
spacy>=3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0.tar.gz
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.0/en_core_web_md-3.7.0.tar.gz
tabulate>=0.8.10
matplotlib>=3.7.1
networkx>=3.1
scikit-learn>=1.3.0
pandas>=2.0.3
numpy>=1.24.3
sqlalchemy>=1.4.0
requests>=2.27.0
neo4j>=5.14.1
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.5
huggingface-hub>=0.19.4
gpt4all>=2.5.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2 